폴더 분리(app.tsx의 일부분 조금 남음)
