+++ src/App.tsx
@@ -1,375 +1,24 @@
-import React, { useState, createContext, useContext } from "react";
-import { generateItems, renderLog } from "./utils";
-import { memo, useCallback, useMemo } from "./@lib";
-
-// 타입 정의
-interface Item {
-  id: number;
-  name: string;
-  category: string;
-  price: number;
-}
-
-interface User {
-  id: number;
-  name: string;
-  email: string;
-}
-
-interface Notification {
-  id: number;
-  message: string;
-  type: "info" | "success" | "warning" | "error";
-}
-
-// AppContext 타입 정의
-interface ThemeContextType {
-  theme: string;
-  toggleTheme: () => void;
-}
-
-interface UserContextType {
-  user: User | null;
-  login: (email: string, password: string) => void;
-  logout: () => void;
-}
-
-interface NotificationContextType {
-  notifications: Notification[];
-  addNotification: (message: string, type: Notification["type"]) => void;
-  removeNotification: (id: number) => void;
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-const UserContext = createContext<UserContextType | undefined>(undefined);
-const NotificationContext = createContext<NotificationContextType | undefined>(
-  undefined
-);
-
-const useThemeContext = () => {
-  const context = useContext(ThemeContext);
-  if (context === undefined) {
-    throw new Error("useThemeContext must be used within an ThemeProvider");
-  }
-  return context;
-};
-
-const useUserContext = () => {
-  const context = useContext(UserContext);
-  if (context === undefined) {
-    throw new Error("useUserContext must be used within an UserProvider");
-  }
-  return context;
-};
-
-const useNotificationContext = () => {
-  const context = useContext(NotificationContext);
-  if (context === undefined) {
-    throw new Error(
-      "useNotificationContext must be used within a NotificationProvider"
-    );
-  }
-  return context;
-};
-
-// Header 컴포넌트
-export const Header: React.FC = memo(() => {
-  renderLog("Header rendered");
-  const { theme, toggleTheme } = useThemeContext();
-  const { user, login, logout } = useUserContext();
-
-  const handleLogin = useCallback(() => {
-    // 실제 애플리케이션에서는 사용자 입력을 받아야 합니다.
-    login("user@example.com", "password");
-  }, [login]);
-
-  return (
-    <header className="bg-gray-800 text-white p-4">
-      <div className="container mx-auto flex justify-between items-center">
-        <h1 className="text-2xl font-bold">샘플 애플리케이션</h1>
-        <div className="flex items-center">
-          <button
-            onClick={toggleTheme}
-            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mr-2"
-          >
-            {theme === "light" ? "다크 모드" : "라이트 모드"}
-          </button>
-          {user ? (
-            <div className="flex items-center">
-              <span className="mr-2">{user.name}님 환영합니다!</span>
-              <button
-                onClick={logout}
-                className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded"
-              >
-                로그아웃
-              </button>
-            </div>
-          ) : (
-            <button
-              onClick={handleLogin}
-              className="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded"
-            >
-              로그인
-            </button>
-          )}
-        </div>
-      </div>
-    </header>
-  );
-});
-
-// ItemList 컴포넌트
-export const ItemList: React.FC<{
-  items: Item[];
-  onAddItemsClick: () => void;
-}> = memo(({ items, onAddItemsClick }) => {
-  renderLog("ItemList rendered");
-  const [filter, setFilter] = useState("");
-  const { theme } = useThemeContext();
-
-  const filteredItems = useMemo(
-    () =>
-      items.filter(
-        (item) =>
-          item.name.toLowerCase().includes(filter.toLowerCase()) ||
-          item.category.toLowerCase().includes(filter.toLowerCase())
-      ),
-    [items]
-  );
-
-  const totalPrice = useMemo(
-    () => filteredItems.reduce((sum, item) => sum + item.price, 0),
-    [filteredItems]
-  );
-
-  const averagePrice = useMemo(
-    () => Math.round(totalPrice / filteredItems.length) || 0,
-    [totalPrice, filteredItems]
-  );
-
-  return (
-    <div className="mt-8">
-      <div className="flex justify-between items-center mb-4">
-        <h2 className="text-2xl font-bold">상품 목록</h2>
-        <div>
-          <button
-            type="button"
-            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded text-xs"
-            onClick={onAddItemsClick}
-          >
-            대량추가
-          </button>
-        </div>
-      </div>
-      <input
-        type="text"
-        placeholder="상품 검색..."
-        value={filter}
-        onChange={(e) => setFilter(e.target.value)}
-        className="w-full p-2 mb-4 border border-gray-300 rounded text-black"
-      />
-      <ul className="mb-4 mx-4 flex gap-3 text-sm justify-end">
-        <li>검색결과: {filteredItems.length.toLocaleString()}개</li>
-        <li>전체가격: {totalPrice.toLocaleString()}원</li>
-        <li>평균가격: {averagePrice.toLocaleString()}원</li>
-      </ul>
-      <ul className="space-y-2">
-        {filteredItems.map((item, index) => (
-          <li
-            key={index}
-            className={`p-2 rounded shadow ${theme === "light" ? "bg-white text-black" : "bg-gray-700 text-white"}`}
-          >
-            {item.name} - {item.category} - {item.price.toLocaleString()}원
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-});
-
-// ComplexForm 컴포넌트
-export const ComplexForm: React.FC = memo(() => {
-  renderLog("ComplexForm rendered");
-  const { addNotification } = useNotificationContext();
-  const [formData, setFormData] = useState(() => ({
-    name: "",
-    email: "",
-    age: 0,
-    preferences: [] as string[],
-  }));
-
-  const handleSubmit = useCallback(
-    (e: React.FormEvent) => {
-      e.preventDefault();
-      addNotification("폼이 성공적으로 제출되었습니다", "success");
-    },
-    [addNotification]
-  );
-
-  const handleInputChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const { name, value } = e.target;
-      setFormData((prev) => ({
-        ...prev,
-        [name]: name === "age" ? parseInt(value) || 0 : value,
-      }));
-    },
-    [setFormData]
-  );
-
-  const handlePreferenceChange = useCallback(
-    (preference: string) => {
-      setFormData((prev) => ({
-        ...prev,
-        preferences: prev.preferences.includes(preference)
-          ? prev.preferences.filter((p) => p !== preference)
-          : [...prev.preferences, preference],
-      }));
-    },
-    [setFormData]
-  );
-
-  return (
-    <div className="mt-8">
-      <h2 className="text-2xl font-bold mb-4">복잡한 폼</h2>
-      <form onSubmit={handleSubmit} className="space-y-4">
-        <input
-          type="text"
-          name="name"
-          value={formData.name}
-          onChange={handleInputChange}
-          placeholder="이름"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <input
-          type="email"
-          name="email"
-          value={formData.email}
-          onChange={handleInputChange}
-          placeholder="이메일"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <input
-          type="number"
-          name="age"
-          value={formData.age}
-          onChange={handleInputChange}
-          placeholder="나이"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <div className="space-x-4">
-          {["독서", "운동", "음악", "여행"].map((pref) => (
-            <label key={pref} className="inline-flex items-center">
-              <input
-                type="checkbox"
-                checked={formData.preferences.includes(pref)}
-                onChange={() => handlePreferenceChange(pref)}
-                className="form-checkbox h-5 w-5 text-blue-600"
-              />
-              <span className="ml-2">{pref}</span>
-            </label>
-          ))}
-        </div>
-        <button
-          type="submit"
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-        >
-          제출
-        </button>
-      </form>
-    </div>
-  );
-}) as React.FC;
-
-// NotificationSystem 컴포넌트
-export const NotificationSystem: React.FC = memo(() => {
-  renderLog("NotificationSystem rendered");
-  const { notifications, removeNotification } = useNotificationContext();
-
-  return (
-    <div className="fixed bottom-4 right-4 space-y-2">
-      {notifications.map((notification) => (
-        <div
-          key={notification.id}
-          className={`p-4 rounded shadow-lg ${
-            notification.type === "success"
-              ? "bg-green-500"
-              : notification.type === "error"
-                ? "bg-red-500"
-                : notification.type === "warning"
-                  ? "bg-yellow-500"
-                  : "bg-blue-500"
-          } text-white`}
-        >
-          {notification.message}
-          <button
-            onClick={() => removeNotification(notification.id)}
-            className="ml-4 text-white hover:text-gray-200"
-          >
-            닫기
-          </button>
-        </div>
-      ))}
-    </div>
-  );
-}) as React.FC;
+import React from "react";
+import { useMemo } from "./@lib";
+import { NotificationContext, ThemeProvider, UserContext } from "./contexts";
+import { Header } from "./components/Header";
+import { ItemList } from "./components/ItemList";
+import { ComplexForm } from "./components/ComplexForm";
+import { NotificationSystem } from "./components/NotificationSystem";
+import { useAuth, useNotification, useTheme, useUser } from "./hooks";
+import { useItems } from "./hooks/useItmes";
 
 // 메인 App 컴포넌트
 const App: React.FC = () => {
-  const [theme, setTheme] = useState("light");
-  const [items, setItems] = useState(() => generateItems(1000));
-  const [user, setUser] = useState<User | null>(null);
-  const [notifications, setNotifications] = useState<Notification[]>([]);
-
-  const toggleTheme = useCallback(() => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  const addItems = useCallback(() => {
-    setItems((prevItems) => [
-      ...prevItems,
-      ...generateItems(1000, prevItems.length),
-    ]);
-  }, []);
-
-  const addNotification = useCallback(
-    (message: string, type: Notification["type"]) => {
-      const newNotification: Notification = {
-        id: Date.now(),
-        message,
-        type,
-      };
-      setNotifications((prev) => [...prev, newNotification]);
-    },
-    []
-  );
-
-  const login = useCallback(
-    (email: string) => {
-      setUser({ id: 1, name: "홍길동", email });
-      addNotification("성공적으로 로그인되었습니다", "success");
-    },
-    [addNotification]
-  );
-
-  const logout = useCallback(() => {
-    setUser(null);
-    addNotification("로그아웃되었습니다", "info");
-  }, [addNotification]);
-
-  const removeNotification = useCallback((id: number) => {
-    setNotifications((prev) =>
-      prev.filter((notification) => notification.id !== id)
-    );
-  }, []);
-
-  const themeContextValue: ThemeContextType = useMemo(
-    () => ({
-      theme,
-      toggleTheme,
-    }),
-    [theme, toggleTheme]
-  );
+  const { theme } = useTheme();
+  const { items, addItems } = useItems();
+  const { user, setUser } = useUser();
+  const { notifications, addNotification, removeNotification } =
+    useNotification();
+  const { login, logout } = useAuth({
+    setUser,
+    addNotification,
+  });
 
   const userContextValue: UserContextType = useMemo(
     () => ({
@@ -390,7 +39,7 @@ const App: React.FC = () => {
   );
 
   return (
-    <ThemeContext.Provider value={themeContextValue}>
+    <ThemeProvider>
       <UserContext.Provider value={userContextValue}>
         <NotificationContext.Provider value={notificationContextValue}>
           <div
@@ -411,7 +60,7 @@ const App: React.FC = () => {
           </div>
         </NotificationContext.Provider>
       </UserContext.Provider>
-    </ThemeContext.Provider>
+    </ThemeProvider>
   );
 };
