+++ src/contexts/ThemeContext.tsx
@@ -1,0 +1,38 @@
+import { createContext, useContext } from "react";
+import { useMemo } from "../@lib";
+import { useTheme } from "../hooks";
+
+interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined
+);
+
+export const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
+  const { theme, toggleTheme } = useTheme();
+
+  const themeContextValue: ThemeContextType = useMemo(
+    () => ({
+      theme,
+      toggleTheme,
+    }),
+    [theme, toggleTheme]
+  );
+
+  return (
+    <ThemeContext.Provider value={themeContextValue}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within an ThemeProvider");
+  }
+  return context;
+};
