+++ src/contexts/UserContext.tsx
@@ -1,0 +1,19 @@
+import { createContext, useContext } from "react";
+import { User } from "../types";
+interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined
+);
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useUserContext must be used within an UserProvider");
+  }
+  return context;
+};
